WIP Tests for projects
