+++ mlflow/projects.py
@@ -309,6 +309,12 @@ def _get_work_dir(uri, use_temp_cwd):
     return uri
 
 
+def _get_storage_dir(storage_dir):
+    if storage_dir is not None and not os.path.exists(storage_dir):
+        os.makedirs(storage_dir)
+    return tempfile.mkdtemp(dir=storage_dir)
+
+
 def _expand_uri(uri):
     if _GIT_URI_REGEX.match(uri):
         return uri
@@ -350,11 +356,31 @@ def _fetch_git_repo(uri, version, dst_dir):
         repo.heads.master.checkout()
 
 
+def _get_conda_env_name(conda_env_path):
+    with open(conda_env_path) as conda_env_file:
+        return hashlib.sha1(conda_env_file.read().encode("utf-8")).hexdigest()
+
+
+def _maybe_create_conda_env(conda_env_path):
+    conda_env = _get_conda_env_name(conda_env_path)
+    (exit_code, _, stderr) = process.exec_cmd(["conda", "--help"], throw_on_error=False)
+    if exit_code != 0:
+        eprint(stderr)
+        raise ExecutionException('conda is not installed properly. Please follow the instructions '
+                                 'on https://conda.io/docs/user-guide/install/index.html')
+    (_, stdout, stderr) = process.exec_cmd(["conda", "env", "list", "--json"])
+    env_names = [os.path.basename(env) for env in json.loads(stdout)['envs']]
+
+    conda_action = 'create'
+    if conda_env not in env_names:
+        eprint('=== Creating conda environment %s ===' % conda_env)
+        process.exec_cmd(["conda", "env", conda_action, "-n", conda_env, "--file",
+                          conda_env_path], stream_output=True)
+
+
 def _run_project(project, entry_point, work_dir, parameters, use_conda, storage_dir, experiment_id):
     """Locally run a project that has been checked out in `work_dir`."""
-    if storage_dir is not None and not os.path.exists(storage_dir):
-        os.makedirs(storage_dir)
-    storage_dir_for_run = tempfile.mkdtemp(dir=storage_dir)
+    storage_dir_for_run = _get_storage_dir(storage_dir)
     eprint("=== Created directory %s for downloading remote URIs passed to arguments of "
            "type 'path' ===" % storage_dir_for_run)
     # Try to build the command first in case the user mis-specified parameters
@@ -362,24 +388,9 @@ def _run_project(project, entry_point, work_dir, parameters, use_conda, storage_
         .compute_command(parameters, storage_dir_for_run)
     commands = []
     if use_conda:
-        with open(os.path.join(work_dir, project.conda_env)) as conda_env_file:
-            conda_env_sha = hashlib.sha1(conda_env_file.read().encode("utf-8")).hexdigest()
-        conda_env = "mlflow-%s" % conda_env_sha
-        (exit_code, _, stderr) = process.exec_cmd(["conda", "--help"], throw_on_error=False)
-        if exit_code != 0:
-            eprint('conda is not installed properly. Please follow the instructions on '
-                   'https://conda.io/docs/user-guide/install/index.html')
-            eprint(stderr)
-            sys.exit(1)
-        (_, stdout, stderr) = process.exec_cmd(["conda", "env", "list", "--json"])
-        env_names = [os.path.basename(env) for env in json.loads(stdout)['envs']]
-
-        conda_action = 'create'
-        if conda_env not in env_names:
-            eprint('=== Creating conda environment %s ===' % conda_env)
-            process.exec_cmd(["conda", "env", conda_action, "-n", conda_env, "--file",
-                              project.conda_env], cwd=work_dir, stream_output=True)
-        commands.append("source activate %s" % conda_env)
+        conda_env_path = os.path.abspath(os.path.join(work_dir, project.conda_env))
+        _maybe_create_conda_env(conda_env_path)
+        commands.append("source activate %s" % _get_conda_env_name(conda_env_path))
 
     # Create a new run and log every provided parameter into it.
     active_run = tracking.start_run(experiment_id=experiment_id,
@@ -387,8 +398,9 @@ def _run_project(project, entry_point, work_dir, parameters, use_conda, storage_
                                     source_version=tracking._get_git_commit(work_dir),
                                     entry_point_name=entry_point,
                                     source_type=SourceType.PROJECT)
-    for key, value in parameters.items():
-        active_run.log_param(Param(key, value))
+    if parameters is not None:
+        for key, value in parameters.items():
+            active_run.log_param(Param(key, value))
     # Add the run id into a magic environment variable that the subprocess will read,
     # causing it to reuse the run.
     exp_id = experiment_id or tracking._get_experiment_id()
