+++ tests/resources/example_project/conda.yaml
@@ -1,0 +1,6 @@
+name: tutorial
+channels:
+  - defaults
+dependencies:
+  - pip:
+    - mlflow
